Move code outside of the destination file
